+++ lib/web_translate_it/string.rb
@@ -62,7 +62,7 @@ module WebTranslateIt
       request.add_field("X-Client-Version", WebTranslateIt::Util.version)
 
       begin
-        response = Util.handle_response(Connection.http_connection.request(request), true)
+        response = Util.handle_response(Connection.http_connection.request(request), true, true)
         strings = []
         YAML.load(response).each do |string_response|
           string = WebTranslateIt::String.new(string_response)
@@ -96,9 +96,8 @@ module WebTranslateIt
       request.add_field("X-Client-Version", WebTranslateIt::Util.version)
 
       begin
-        response = Connection.http_connection.request(request)
-        return nil if response.code.to_i == 404
-        string = WebTranslateIt::String.new(YAML.load(response.body))
+        response = Util.handle_response(Connection.http_connection.request(request), true, true)
+        string = WebTranslateIt::String.new(YAML.load(response))
         string.new_record = false
         return string
       rescue Timeout::Error
@@ -139,7 +138,7 @@ module WebTranslateIt
       request.add_field("X-Client-Version", WebTranslateIt::Util.version)
 
       begin
-        Util.handle_response(Connection.http_connection.request(request), true)
+        Util.handle_response(Connection.http_connection.request(request), true, true)
       rescue Timeout::Error
         puts "The request timed out. The service may be overloaded. We will retry in 5 seconds."
         sleep(5)
@@ -164,7 +163,7 @@ module WebTranslateIt
       request.add_field("X-Client-Version", WebTranslateIt::Util.version)
 
       begin
-        response = Util.handle_response(Connection.http_connection.request(request), true)
+        response = Util.handle_response(Connection.http_connection.request(request), true, true)
         hash = YAML.load(response)
         return nil if hash.empty?
         translation = WebTranslateIt::Translation.new(hash)
@@ -195,7 +194,7 @@ module WebTranslateIt
       end
       
       begin
-        Util.handle_response(Connection.http_connection.request(request), true)
+        Util.handle_response(Connection.http_connection.request(request), true, true)
       rescue Timeout::Error
         puts "The request timed out. The service may be overloaded. We will retry in 5 seconds."
         sleep(5)
@@ -214,7 +213,7 @@ module WebTranslateIt
       request.body = self.to_json(true)
 
       begin
-        response = YAML.load(Util.handle_response(Connection.http_connection.request(request), true))
+        response = YAML.load(Util.handle_response(Connection.http_connection.request(request), true, true))
         self.id = response["id"]
         self.new_record = false
         return true