+++ lib/web_translate_it/translation.rb
@@ -49,7 +49,7 @@ module WebTranslateIt
       request.body = self.to_hash.to_json
 
       begin
-        Util.handle_response(Connection.http_connection.request(request), true)
+        Util.handle_response(Connection.http_connection.request(request), true, true)
       rescue Timeout::Error
         puts "The request timed out. The service may be overloaded. We will retry in 5 seconds."
         sleep(5)