+++ history.md
@@ -1,3 +1,7 @@
+## Edge
+
+* String, Translation, Term and TermTranslation classes now raise exceptions and display useful error messages. #87
+
 ## Version 2.0.1 / 2012-05-11
 
 * Fix: Issue with saving labels on a string.